changes in project structure and html codes. changed app.py according to project structure.
