--- codestyles/Default.xml
@@ -1,3 +1,0 @@
-<code_scheme name="Default" version="173">
-  <option name="ALIGN_MULTILINE_PARAMETERS_IN_CALLS" value="true" />
-</code_scheme>
