--- editor.codeinsight.xml
@@ -1,6 +1,0 @@
-<application>
-  <component name="CodeInsightSettings">
-    <option name="REFORMAT_ON_PASTE" value="1" />
-    <option name="INDENT_TO_CARET_ON_PASTE" value="true" />
-  </component>
-</application>
