+++ manage.py
@@ -1,0 +1,10 @@
+
+
+
+
+
+
+
+if __name__ == '__main__':
+
+    import billing_portal.app
