--- project.default.xml
@@ -1,10 +1,0 @@
-<application>
-  <component name="ProjectManager">
-    <defaultProject>
-      <component name="InspectionProjectProfileManager">
-        <option name="PROJECT_PROFILE" />
-        <version value="1.0" />
-      </component>
-    </defaultProject>
-  </component>
-</application>
