--- templates/base.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>The United Cargo</title>
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <script type="text/javascript" src="{{ url_for('static', filename='js/jquery-1.12.4.min.js')}}"></script>
-     <script type="text/javascript" src="{{ url_for('static', filename='js/bootstrap.min.js')}}"></script>
- <script type="text/javascript" src="{{ url_for('static', filename='js/myscript.js')}}"></script>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css')}}"/>
-    <style>
-        body {
-        background-color: #eee;
-        margin-top: 60px;
-        }
-    </style>
-</head>
-<body>
-
-
-	<div class="container" style="min-height:100% width:80%">
-	{% with messages = get_flashed_messages() %}
-	  {% if messages %}
-	    {% for message in messages %}
-		  <div class="alert alert-warning alert-dismissible" role="alert">
-		  <span class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></span>
-			{{message}}
-		  </div>
-		{% endfor %}
-	  {% endif %}
-	{% endwith %}
-
-        {% block page_content %}
-
-{% endblock %}
-
-	</div>
-
-
-
-</body>
-</html>
