--- templates/fields.html
@@ -1,199 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>The United Cargo</title>
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <script  src="{{ url_for('static', filename='vendor/jquery/jquery.min.js')}}"></script>
-     <script src="{{ url_for('static', filename='js/bootstrap.min.js')}}"></script>
- <script  src="{{ url_for('static', filename='js/myscript.js')}}"></script>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css')}}"/>
-    <style>
-        body {
-        background-color: #eee;
-        margin-top: 60px;
-        }
-    </style>
-</head>
-<body>
-
-
-
-  <script type="text/javascript">
-
-
-
-$(document).ready(function(){
-    $("#datepicker").datepicker();
-});
-
-
-</script>
-
-
-
-
-<nav class="navbar navbar-inverse navbar-fixed-top">
-      <div class="container-fluid">
-        <div class="navbar-header">
-          <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-            <span class="sr-only">Toggle navigation</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-          </button>
-          <a class="navbar-brand" href="#">The United Cargo</a>
-        </div>
-        <div id="navbar" class="navbar-collapse collapse">
-          <ul class="nav navbar-nav navbar-right">
-            <li><a href="#">Dashboard</a></li>
-            <li><a href="#">Settings</a></li>
-            <li><a href="#">Profile</a></li>
-            <li><a href={{url_for('home_template')}}>Log Out</a></li>
-          </ul>
-          <form class="navbar-form navbar-right">
-            <input type="text" class="form-control" placeholder="Search...">
-          </form>
-        </div>
-      </div>
-    </nav>
-
-
-
-<div class="container-fluid" style="margin-top: 40px;">
-    <div class="row">
-
-
-
-            <h2 id="fieldsheading" style="font-family:times, serif; margin-bottom:40px;" class="text-center"><u>Add New Entry</u></h2>
-
-
-
-            <div>
-                <form action="{{url_for('handle_data')}}" method="post" enctype="multipart/form-data">
-                    <fieldset>
-                        <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="date"><span class="FieldInfo">Date:</span> </label>
-                            <input class="form-control" type="date" name="Date" id="Date" placeholder="Date">
-                        </div>
-                        </div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="grno"><span class="FieldInfo">Gr.No:</span> </label>
-                            <input class="form-control" type="text" name="GrNo" id="GrNo" placeholder="GrNo">
-                        </div>
-                         </div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="pkgs"><span class="FieldInfo">Pkgs:</span> </label>
-                            <input class="form-control" type="text" name="Pkgs" id="Pkgs" placeholder="Pkgs">
-                        </div>
-                         </div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="awt"><span class="FieldInfo">A WT:</span> </label>
-                            <input class="form-control" type="text" name="Awt" id="awt" placeholder="awt">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="cwt"><span class="FieldInfo">C WT:</span> </label>
-                            <input class="form-control" type="text" name="Cwt" id="cwt" placeholder="cwt">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="invoice"><span class="FieldInfo">Invoice No:</span> </label>
-                            <input class="form-control" type="text" name="Invoice" id="Invoice" placeholder="Invoice No">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="sender"><span class="FieldInfo">Sender:</span> </label>
-                            <input class="form-control" type="text" name="Sender" id="Sender" placeholder="Sender">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="receiver"><span class="FieldInfo">Receiver:</span> </label>
-                            <input class="form-control" type="text" name="Receiver" id="Receiver" placeholder="Receiver">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="origin"><span class="FieldInfo">Origin:</span> </label>
-                            <input class="form-control" type="text" name="Origin" id="Origin" placeholder="Origin">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="destination"><span class="FieldInfo">Destination:</span> </label>
-                            <input class="form-control" type="text" name="Destination" id="Destination" placeholder="Destination">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="mode"><span class="FieldInfo">Mode:</span> </label>
-                            <input class="form-control" type="text" name="Mode" id="Mode" placeholder="Mode">
-                        </div></div>
-
-                         <div class="col-sm-2">
-                        <div class="form-group">
-                            <label for="fright"><span class="FieldInfo">Freight:</span> </label>
-                            <input class="form-control" type="text" name="Freight" id="Fright" placeholder="Freight">
-                        </div></div>
-
-
-
-                         <div class="col-sm-2 col-sm-offset-5" style="margin-top: 40px;">
-                        <input class="btn btn-success btn-block " type="Submit" name="Submit" value="Add New Entry">
-                         </div>
-                             <br>
-                    </fieldset>
-
-                </form>
-            </div>
-
-
-
-
-
-
-
-
-
-
-
-
-        <!-- Ending of main area-->
-
-    </div> <!-- Ending of Row-->
-</div> <!-- Ending of Container-->
-
-<div id="Footer" style="padding: 10px;
-	margin-top: 70px;
-	border-top: 1px solid black;
-	color: #eeeeee;
-	background-color: #211f22;
-	text-align: center;">
-
-
-    <hr>
-    <p style="text-align: center; color: #ffffff;">United Cargo | &copy; 2017-2018</p>
-    <p style="text-align: center; color: #ffffff;">Door to Door Cargo Air and Train Services</p>
-    <p style="text-align: center; color: #ffffff;">Sr.No. 16/1D/2K/1, Behind Durga Hotel, Wadgaon Shinde Road, Lohegaon, Pune-411047</p>
-    <p style="text-align: center; color: #ffffff;">Ph.: 9373344474, 9371344474 E-mail: singh.neeraj02@gmail.com</p>
-
-    <hr>
-</div>
-
-<div style="height: 10px; background: #27aae1;"></div>
-
-
-
-</body>
-</html>
