--- vim_settings.xml
@@ -1,15 +1,0 @@
-<application>
-  <component name="VimSettings">
-    <state version="4" enabled="true" />
-    <globalmarks />
-    <filemarks />
-    <jumps />
-    <registers />
-    <search>
-      <last-dir>0</last-dir>
-      <show-last>false</show-last>
-    </search>
-    <history />
-    <shortcut-conflicts />
-  </component>
-</application>
